refactor: add layout inteface, separate IT layout
